requirements.txt
python-dotenv
passlib
cryptography

alembic
psycopg[binary]
sqlalchemy

pydantic
uvicorn
fastapi
# python-multipart

black